requirements.txt
wheel
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
Werkzeug==1.0.1
pymysql==0.9.3
SQLAlchemy-serializer==1.3.4.2
uWSGI
# flask-swagger
# swagger-ui
# flask-swagger-ui
pytest
mock
